// src/api/growthapiComponents.ts
export interface Goal {
	id: string
	title: string
	description: string
	category: string
	dueDate?: string
	progress: number
	completed: boolean
	relatedHabitIds?: Array<string>
	userId: string
	createdAt: string
	updatedAt: string
}